+++ R/zzz.R
@@ -3,13 +3,24 @@
   
   install_cgal <- isTRUE(Sys.getenv("CGAL_DOWNLOAD")  != "0")
   interact     <- interactive()
+  no_cgal      <- !cgal_is_installed()
+  
+  # previous CRAN requests to not include header files need this function
+  .cgal_download_check() # will see if interactive and ask to download
+  
+}
+
+.depricated_onload <- function(libname, pkgname) {
+  
+  install_cgal <- isTRUE(Sys.getenv("CGAL_DOWNLOAD")  != "0")
+  interact     <- interactive()
   # on_cran      <- !identical(Sys.getenv("NOT_CRAN"), "true")
-  has_internet <- curl::has_internet()
+  # has_internet <- curl::has_internet()
   no_cgal      <- !cgal_is_installed()
   default_url  <- cgal_pkg_state$DEFAULT_URL
   default_vers <- cgal_pkg_state$DEFAULT_VERSION
   
-  if(no_cgal && has_internet && install_cgal && !interact ) {
+  if(no_cgal && install_cgal && !interact ) {
     # to comply with CRAN, must download a fixed version of files by default.
     # this will check for environment variable otherwise will pass default
     if (Sys.getenv("CGAL_DIR") != "") {
@@ -18,7 +29,7 @@
     
     tryCatch(suppressMessages(cgal_install(cgal_path = default_url)),
              error = .cgal_install_error_catch
-             )
+    )
     
     if (Sys.getenv("CGAL_DIR") == "") {
       # needs another check currently to save proper version
@@ -27,6 +38,7 @@
     
   }
   
+  # previous CRAN requests to not include header files need this function
   .cgal_download_check() # will see if interactive and ask to download
   
 }