+++ tools/config/cleanup.R
@@ -1,0 +1,5 @@
+# Clean up files generated during configuration here.
+# Use 'remove_file()' to remove files generated during configuration.
+
+remove_file(file.path("inst","include","CGAL"))
+remove_file(file.path("include","CGAL_zip.tar.xz"))