+++ vignettes/Installation.R
@@ -1,0 +1,15 @@
+## ---- eval = FALSE------------------------------------------------------------
+#  install.packages("RcppCGAL")
+
+## ---- eval = FALSE------------------------------------------------------------
+#  Sys.setenv(CGAL_DIR="https://some/url/cgalheaders.tar.gz")
+
+## ---- eval = FALSE------------------------------------------------------------
+#  set_cgal("https://some/url/cgalheaders.tar.gz")
+
+## ---- eval = FALSE------------------------------------------------------------
+#  Sys.setenv(CGAL_DIR="path/to/include/CGAL")
+
+## ---- eval = FALSE------------------------------------------------------------
+#  set_cgal("path/to/include/CGAL)
+